requirements.txt
pandas>=2.2.0
python-binance>=1.0.36
python-dotenv>=1.0.0
